# docs/archive/product-history/LobsterPark_Docs_V1_4/LobsterPark_OpenAPI_V1_4.yaml
openapi: 3.1.0
info:
  title: Lobster Park API
  version: 1.4.0
  description: >
    OpenAPI contract for Lobster Park. V1.4 adds Secrets CRUD endpoints,
    enhances alerts/notifications filters, fixes PatchInstanceRequest.name
    maxLength, and adds quotaJson to Tenant schema.
servers:
- url: https://lobster-park.internal
security:
- cookieAuth: []
paths:
...
  /api/v1/nodes/pairing-requests:
    get:
      summary: List pairing requests
      operationId: listPairingRequests
      parameters:
      - $ref: '#/components/parameters/PageNo'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePagedPairingRequests'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/nodes/pairing-requests/{requestId}/approve:
    post:
      summary: Approve pairing request
      operationId: approvePairingRequest
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePairingRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/nodes/pairing-requests/{requestId}/reject:
    post:
      summary: Reject pairing request
      operationId: rejectPairingRequest
      parameters:
      - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePairingRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/v1/instances/{instanceId}/openclaw/basic-config:
    get:
      summary: Get instance OpenClaw basic config
      operationId: getOpenClawBasicConfig
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawBasicConfig'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    put:
      summary: Save instance OpenClaw basic config into draft
      operationId: saveOpenClawBasicConfig
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenClawBasicConfigInput'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawBasicConfig'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/channels:
    get:
      summary: List OpenClaw channels for instance
      operationId: listOpenClawInstanceChannels
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawInstanceChannels'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/channels/{channelType}/connect:
    post:
      summary: Connect or update an OpenClaw channel
      operationId: connectOpenClawChannel
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/ChannelType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenClawChannelConnectRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawChannelConnectResult'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/channels/{channelType}/test:
    post:
      summary: Test an OpenClaw channel connection
      operationId: testOpenClawChannel
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/ChannelType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenClawChannelTestRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawChannelTestResult'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/channels/{channelType}/runtime-status:
    get:
      summary: Get runtime session status for an OpenClaw channel
      operationId: getOpenClawChannelRuntimeStatus
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/ChannelType'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawChannelRuntimeStatus'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/channels/{channelType}/logs:
    get:
      summary: Get recent OpenClaw channel logs for an instance channel
      operationId: getOpenClawChannelLogs
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/ChannelType'
      - in: query
        name: lines
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 500
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawChannelLogs'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/channels/{channelType}/qr-diagnostics:
    get:
      summary: Get QR/session diagnostics for a QR-based OpenClaw channel
      operationId: getOpenClawQrDiagnostics
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/ChannelType'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawQrDiagnostics'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/channels/{channelType}/qr-session/start:
    post:
      summary: Start QR login session for a QR-based OpenClaw channel
      operationId: startOpenClawQrSession
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/ChannelType'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawQrSession'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/channels/{channelType}/qr-session/wait:
    get:
      summary: Wait for QR login completion for a QR-based OpenClaw channel
      operationId: waitOpenClawQrSession
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/ChannelType'
      - in: query
        name: timeoutMs
        required: false
        schema:
          type: integer
          minimum: 1000
          maximum: 60000
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawQrSession'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/pairing-requests:
    get:
      summary: List instance OpenClaw pairing requests
      operationId: listOpenClawInstancePairingRequests
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawPairingRequests'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/pairing-requests/{pairingCode}/approve:
    post:
      summary: Approve instance OpenClaw pairing request
      operationId: approveOpenClawInstancePairingRequest
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/PairingCode'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePairingRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/pairing-requests/{pairingCode}/reject:
    post:
      summary: Reject instance OpenClaw pairing request
      operationId: rejectOpenClawInstancePairingRequest
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/PairingCode'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenClawPairingRejectRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePairingRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/console/session:
    post:
      summary: Create instance OpenClaw console session snapshot
      operationId: createOpenClawConsoleSession
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawConsoleSession'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/console/send:
    post:
      summary: Send a message through instance OpenClaw console
      operationId: sendOpenClawConsoleMessage
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawConsoleSession'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/openclaw/console/history:
    get:
      summary: Get recent instance OpenClaw console history
      operationId: getOpenClawConsoleHistory
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 50
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawConsoleHistory'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/v1/catalog/channels:
    get:
      summary: List platform-supported OpenClaw channels
      operationId: listOpenClawCatalogChannels
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawCatalogChannels'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/catalog/channel-plugins:
    get:
      summary: List platform-supported OpenClaw channel plugins
      operationId: listOpenClawCatalogChannelPlugins
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOpenClawCatalogChannelPlugins'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: lp_access
  parameters:
    PageNo:
      in: query
      name: pageNo
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    InstanceId:
      in: path
      name: instanceId
      required: true
      schema:
        type: string
...
    RequestId:
      in: path
      name: requestId
      required: true
      schema:
        type: string
...
    ChannelType:
      in: path
      name: channelType
      required: true
      schema:
        type: string
    PairingCode:
      in: path
      name: pairingCode
      required: true
      schema:
        type: string
  schemas:
    EnvelopeBase:
      type: object
      properties:
        requestId:
          type: string
        code:
          type: integer
        message:
          type: string
    PagedMeta:
      type: object
      properties:
        pageNo:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    PairingStatus:
      type: string
      enum:
      - pending
      - approved
      - rejected
      - expired
...
    PairingRequest:
      type: object
      properties:
        id:
          type: string
        instanceId:
          type: string
        pairingStatus:
          $ref: '#/components/schemas/PairingStatus'
        requestedAt:
          type: string
          format: date-time
        tenantId:
          type: string
        nodeFingerprint:
          type: string
        reviewedBy:
          type: string
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        reason:
          type: string
          nullable: true
...
    OpenClawBasicConfigInput:
      type: object
      properties:
        general:
          type: object
          additionalProperties: true
        defaultModel:
          type: object
          additionalProperties: true
          nullable: true
        defaultAgent:
          type: object
          additionalProperties: true
          nullable: true
        security:
          type: object
          additionalProperties: true
        toolPolicy:
          type: object
          additionalProperties: true
        limits:
          type: object
          additionalProperties: true
        channelDefaults:
          type: object
          additionalProperties: true
    OpenClawBasicConfig:
      type: object
      properties:
        instanceId:
          type: string
        runtimeVersion:
          type: string
        draftDirty:
          type: boolean
        activeVersionId:
          type: string
          nullable: true
        general:
          type: object
          additionalProperties: true
        defaultModel:
          type: object
          additionalProperties: true
          nullable: true
        defaultAgent:
          type: object
          additionalProperties: true
          nullable: true
        security:
          type: object
          additionalProperties: true
        toolPolicy:
          type: object
          additionalProperties: true
        limits:
          type: object
          additionalProperties: true
        channelDefaults:
          type: object
          additionalProperties: true
        channelCount:
          type: integer
        skillCount:
          type: integer
...
    OpenClawCatalogChannels:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OpenClawChannelCatalogItem'
...
    OpenClawInstanceChannels:
      type: object
      properties:
        instanceId:
          type: string
        connectivity:
          $ref: '#/components/schemas/OpenClawConnectivitySummary'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OpenClawInstanceChannel'
    OpenClawChannelConnectRequest:
      type: object
      properties:
        channelId:
          type: string
        modelId:
          type: string
        enabled:
          type: boolean
        testTarget:
          type: string
        fields:
          type: object
          additionalProperties: true
      additionalProperties: true
    OpenClawChannelConnectResult:
      type: object
      properties:
        instanceId:
          type: string
        channelType:
          type: string
        configured:
          type: boolean
        publishRequired:
          type: boolean
        config:
          $ref: '#/components/schemas/OpenClawChannelConfig'
          nullable: true
    OpenClawChannelTestRequest:
      type: object
      properties:
        target:
          type: string
        message:
          type: string
        realDelivery:
          type: boolean
          description: When true, send through the running instance gateway instead of CLI dry-run.
      additionalProperties: true
    OpenClawChannelTestResult:
      type: object
      properties:
        instanceId:
          type: string
        channelType:
          type: string
        success:
          type: boolean
        simulated:
          type: boolean
        testMode:
          type: string
        deliveryMode:
          type: string
          enum:
          - dry_run
          - real
        relayMode:
          type: string
          enum:
          - gateway
          - local_config
        target:
          type: string
        message:
          type: string
        errorMessage:
          type: string
          nullable: true
        relay:
          type: object
          additionalProperties: true
          nullable: true
        checkedAt:
          type: string
          format: date-time
    OpenClawChannelRuntimeStatus:
      type: object
      properties:
        instanceId:
          type: string
        channelType:
          type: string
        connectionMode:
          type: string
        configured:
          type: boolean
        linked:
          type: boolean
        running:
          type: boolean
        connected:
          type: boolean
        sessionStatus:
          type: string
        lastError:
          type: string
          nullable: true
        self:
          type: object
          additionalProperties: true
          nullable: true
        accounts:
          type: array
          items:
            type: object
            additionalProperties: true
        statusSource:
          type: string
        qrSupported:
          type: boolean
        qrHint:
          type: string
          nullable: true
    OpenClawQrSession:
      type: object
      properties:
        instanceId:
          type: string
        channelType:
          type: string
        qrSupported:
          type: boolean
        status:
          type: string
        connected:
          type: boolean
          nullable: true
        qrDataUrl:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        source:
          type: string
    OpenClawChannelLogs:
      type: object
      properties:
        instanceId:
          type: string
        channelType:
          type: string
        lines:
          type: array
          items:
            type: string
        file:
          type: string
          nullable: true
        source:
          type: string
        errorMessage:
          type: string
          nullable: true
    OpenClawQrDiagnostics:
      type: object
      properties:
        instanceId:
          type: string
        channelType:
          type: string
        runtimeStatus:
          type: object
          additionalProperties: true
        logs:
          type: object
          additionalProperties: true
        reasons:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
    OpenClawPairingRequests:
      type: object
      properties:
        instanceId:
          type: string
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/PairingRequest'
    OpenClawPairingRejectRequest:
      type: object
      properties:
        reason:
          type: string
          nullable: true
...
    EnvelopeError:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            type: object
            additionalProperties: true
...
    EnvelopePagedPairingRequests:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            allOf:
            - $ref: '#/components/schemas/PagedMeta'
            - type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/PairingRequest'
    EnvelopePairingRequest:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/PairingRequest'
...
    EnvelopeOpenClawBasicConfig:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OpenClawBasicConfig'
    EnvelopeOpenClawCatalogChannels:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OpenClawCatalogChannels'
    EnvelopeOpenClawCatalogChannelPlugins:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OpenClawCatalogChannelPlugins'
    EnvelopeOpenClawInstanceChannels:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OpenClawInstanceChannels'
    EnvelopeOpenClawChannelConnectResult:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OpenClawChannelConnectResult'
    EnvelopeOpenClawChannelTestResult:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OpenClawChannelTestResult'
    EnvelopeOpenClawChannelRuntimeStatus:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OpenClawChannelRuntimeStatus'
    EnvelopeOpenClawQrSession:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OpenClawQrSession'
    EnvelopeOpenClawChannelLogs:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OpenClawChannelLogs'
    EnvelopeOpenClawQrDiagnostics:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OpenClawQrDiagnostics'
    EnvelopeOpenClawPairingRequests:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OpenClawPairingRequests'
    EnvelopeOpenClawConsoleSession:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OpenClawConsoleSession'
    EnvelopeOpenClawConsoleHistory:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/OpenClawConsoleHistory'
...
  responses:
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnvelopeError'
...
  headers:
    RetryAfter:
      description: Seconds until the client may retry
      schema:
        type: integer